refactor(infrastructure-services): Inline repositories into service
Inline the infrastructure service repositories into the
`InfrastructureServiceService` because they introduced a layer of
abstraction that was not required.

Previsouly, many functions of the `InfrastructureServiceService` were
not tested. To address that, migrate most of the removed tests for the
repositories to the integration test of the service.
