+++ components/infrastructure-services/backend/src/test/kotlin/InfrastructureServiceServiceIntegrationTest.kt
@@ -21,23 +21,40 @@ package org.eclipse.apoapsis.ortserver.components.infrastructureservices
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.WordSpec
+import io.kotest.matchers.collections.shouldContainExactly
+import io.kotest.matchers.collections.shouldContainExactlyInAnyOrder
+import io.kotest.matchers.collections.shouldContainOnly
 import io.kotest.matchers.nulls.beNull
 import io.kotest.matchers.should
 import io.kotest.matchers.shouldBe
+import io.kotest.matchers.shouldNotBe
 import io.kotest.matchers.string.shouldContain
+import io.kotest.matchers.string.shouldInclude
 
 import java.util.EnumSet
 
 import org.eclipse.apoapsis.ortserver.components.secrets.SecretService
+import org.eclipse.apoapsis.ortserver.dao.UniqueConstraintException
 import org.eclipse.apoapsis.ortserver.dao.test.DatabaseTestExtension
+import org.eclipse.apoapsis.ortserver.dao.test.Fixtures
 import org.eclipse.apoapsis.ortserver.model.CredentialsType
+import org.eclipse.apoapsis.ortserver.model.Hierarchy
 import org.eclipse.apoapsis.ortserver.model.InfrastructureService
+import org.eclipse.apoapsis.ortserver.model.InfrastructureServiceDeclaration
 import org.eclipse.apoapsis.ortserver.model.Organization
 import org.eclipse.apoapsis.ortserver.model.OrganizationId
+import org.eclipse.apoapsis.ortserver.model.Product
+import org.eclipse.apoapsis.ortserver.model.ProductId
+import org.eclipse.apoapsis.ortserver.model.Repository
+import org.eclipse.apoapsis.ortserver.model.RepositoryId
+import org.eclipse.apoapsis.ortserver.model.Secret
 import org.eclipse.apoapsis.ortserver.model.util.ListQueryParameters
 import org.eclipse.apoapsis.ortserver.model.util.ListQueryResult
 import org.eclipse.apoapsis.ortserver.model.util.OptionalValue
+import org.eclipse.apoapsis.ortserver.model.util.OrderDirection
+import org.eclipse.apoapsis.ortserver.model.util.OrderField
 import org.eclipse.apoapsis.ortserver.model.util.asPresent
+import org.eclipse.apoapsis.ortserver.model.validation.ValidationException
 import org.eclipse.apoapsis.ortserver.secrets.SecretStorage
 import org.eclipse.apoapsis.ortserver.secrets.SecretsProviderFactoryForTesting
 
@@ -48,9 +65,91 @@ private const val SERVICE_DESC = "This is a test infrastructure service"
 class InfrastructureServiceServiceIntegrationTest : WordSpec({
     val dbExtension = extension(DatabaseTestExtension())
 
+    lateinit var fixtures: Fixtures
     lateinit var organization: Organization
+    lateinit var product: Product
+    lateinit var repository: Repository
+    lateinit var orgUserSecret: Secret
+    lateinit var orgPassSecret: Secret
+    lateinit var prodUserSecret: Secret
+    lateinit var prodPassSecret: Secret
+    lateinit var repoUserSecret: Secret
+    lateinit var repoPassSecret: Secret
     lateinit var secretService: SecretService
     lateinit var service: InfrastructureServiceService
+
+    suspend fun createOrganizationSecret(name: String, organizationId: Long = organization.id) =
+        secretService.createSecret(name, "value", null, OrganizationId(organizationId))
+
+    suspend fun createProductSecret(name: String, productId: Long = product.id) =
+        secretService.createSecret(name, "value", null, ProductId(productId))
+
+    suspend fun createRepositorySecret(name: String, repositoryId: Long = repository.id) =
+        secretService.createSecret(name, "value", null, RepositoryId(repositoryId))
+
+    fun createInfrastructureService(
+        name: String = SERVICE_NAME,
+        url: String = SERVICE_URL,
+        description: String? = SERVICE_DESC,
+        usernameSecret: Secret? = null,
+        passwordSecret: Secret? = null,
+        organization: Organization? = null,
+        product: Product? = null,
+        repository: Repository? = null,
+        credentialsTypes: Set<CredentialsType> = setOf(CredentialsType.NETRC_FILE),
+    ): InfrastructureService =
+        InfrastructureService(
+            name,
+            url,
+            description,
+            usernameSecret ?: when {
+                organization != null -> orgUserSecret
+                product != null -> prodUserSecret
+                repository != null -> repoUserSecret
+                else -> error("At least one of organization, product or repository must be set.")
+            },
+            passwordSecret ?: when {
+                organization != null -> orgPassSecret
+                product != null -> prodPassSecret
+                repository != null -> repoPassSecret
+                else -> error("At least one of organization, product or repository must be set.")
+            },
+            organization,
+            product,
+            repository,
+            credentialsTypes
+        )
+
+    suspend fun createInfrastructureService(infrastructureService: InfrastructureService): InfrastructureService =
+        service.createForId(
+            id = infrastructureService.organization?.let { OrganizationId(it.id) }
+                ?: infrastructureService.product?.let { ProductId(it.id) }
+                ?: infrastructureService.repository?.let { RepositoryId(it.id) }
+                ?: error("At least one of organization, product or repository must be set."),
+            name = infrastructureService.name,
+            url = infrastructureService.url,
+            description = infrastructureService.description,
+            usernameSecretRef = infrastructureService.usernameSecret.name,
+            passwordSecretRef = infrastructureService.passwordSecret.name,
+            credentialsTypes = infrastructureService.credentialsTypes,
+        )
+
+    fun createInfrastructureServiceDeclaration(
+        name: String = SERVICE_NAME,
+        url: String = SERVICE_URL,
+        description: String = SERVICE_DESC,
+        usernameSecret: Secret = orgUserSecret,
+        passwordSecret: Secret = orgPassSecret,
+        credentialsTypes: Set<CredentialsType> = setOf(CredentialsType.NETRC_FILE),
+    ): InfrastructureServiceDeclaration =
+        InfrastructureServiceDeclaration(
+            name,
+            url,
+            description,
+            usernameSecret.name,
+            passwordSecret.name,
+            credentialsTypes
+        )
 
     beforeEach {
         secretService = SecretService(
@@ -61,24 +160,28 @@ class InfrastructureServiceServiceIntegrationTest : WordSpec({
 
         service = InfrastructureServiceService(dbExtension.db, secretService)
 
+        fixtures = dbExtension.fixtures
         organization = dbExtension.fixtures.organization
-    }
+        product = dbExtension.fixtures.product
+        repository = dbExtension.fixtures.repository
 
-    suspend fun createOrganizationSecret(name: String) =
-        secretService.createSecret(name, "value", null, OrganizationId(organization.id))
+        orgUserSecret = createOrganizationSecret("user")
+        orgPassSecret = createOrganizationSecret("pass")
+        prodUserSecret = createProductSecret("user")
+        prodPassSecret = createProductSecret("pass")
+        repoUserSecret = createRepositorySecret("user")
+        repoPassSecret = createRepositorySecret("pass")
+    }
 
     "createForId" should {
         "create an infrastructure service for an organization" {
-            val userSecret = createOrganizationSecret("user")
-            val passSecret = createOrganizationSecret("pass")
-
             val createResult = service.createForId(
                 OrganizationId(organization.id),
                 SERVICE_NAME,
                 SERVICE_URL,
                 SERVICE_DESC,
-                userSecret.name,
-                passSecret.name,
+                orgUserSecret.name,
+                orgPassSecret.name,
                 EnumSet.of(CredentialsType.NETRC_FILE)
             )
 
@@ -86,8 +189,8 @@ class InfrastructureServiceServiceIntegrationTest : WordSpec({
                 SERVICE_NAME,
                 SERVICE_URL,
                 SERVICE_DESC,
-                userSecret,
-                passSecret,
+                orgUserSecret,
+                orgPassSecret,
                 organization,
                 null,
                 null,
@@ -95,29 +198,73 @@ class InfrastructureServiceServiceIntegrationTest : WordSpec({
             )
         }
 
-        "throw an exception if a secret reference cannot be resolved" {
-            val userSecret = createOrganizationSecret("user")
+        "create an infrastructure service for a product" {
+            val createResult = service.createForId(
+                ProductId(product.id),
+                SERVICE_NAME,
+                SERVICE_URL,
+                SERVICE_DESC,
+                prodUserSecret.name,
+                prodPassSecret.name,
+                EnumSet.of(CredentialsType.NETRC_FILE)
+            )
 
+            createResult shouldBe InfrastructureService(
+                SERVICE_NAME,
+                SERVICE_URL,
+                SERVICE_DESC,
+                prodUserSecret,
+                prodPassSecret,
+                null,
+                product,
+                null,
+                EnumSet.of(CredentialsType.NETRC_FILE)
+            )
+        }
+
+        "create an infrastructure service for a repository" {
+            val createResult = service.createForId(
+                RepositoryId(repository.id),
+                SERVICE_NAME,
+                SERVICE_URL,
+                SERVICE_DESC,
+                repoUserSecret.name,
+                repoPassSecret.name,
+                EnumSet.of(CredentialsType.NETRC_FILE)
+            )
+
+            createResult shouldBe InfrastructureService(
+                SERVICE_NAME,
+                SERVICE_URL,
+                SERVICE_DESC,
+                repoUserSecret,
+                repoPassSecret,
+                null,
+                null,
+                repository,
+                EnumSet.of(CredentialsType.NETRC_FILE)
+            )
+        }
+
+        "throw an exception if a secret reference cannot be resolved" {
             val exception = shouldThrow<InvalidSecretReferenceException> {
                 service.createForId(
                     OrganizationId(organization.id),
                     SERVICE_NAME,
                     SERVICE_URL,
                     SERVICE_DESC,
-                    userSecret.name,
-                    "pass",
+                    orgUserSecret.name,
+                    "non-existent",
                     emptySet()
                 )
             }
 
-            exception.message shouldContain "pass"
+            exception.message shouldContain "non-existent"
         }
     }
 
     "updateForId" should {
         "update the properties of a service" {
-            val userSecret = createOrganizationSecret("user")
-            val passSecret = createOrganizationSecret("pass")
             val newUserSecret = createOrganizationSecret("newUser")
             val newpassSecret = createOrganizationSecret("newPass")
 
@@ -126,8 +273,8 @@ class InfrastructureServiceServiceIntegrationTest : WordSpec({
                 SERVICE_NAME,
                 SERVICE_URL,
                 SERVICE_DESC,
-                userSecret.name,
-                passSecret.name,
+                orgUserSecret.name,
+                orgPassSecret.name,
                 EnumSet.of(CredentialsType.NETRC_FILE)
             )
 
@@ -171,16 +318,13 @@ class InfrastructureServiceServiceIntegrationTest : WordSpec({
 
     "deleteForId" should {
         "delete an infrastructure service" {
-            val userSecret = createOrganizationSecret("user")
-            val passSecret = createOrganizationSecret("pass")
-
             service.createForId(
                 OrganizationId(organization.id),
                 SERVICE_NAME,
                 SERVICE_URL,
                 SERVICE_DESC,
-                userSecret.name,
-                passSecret.name,
+                orgUserSecret.name,
+                orgPassSecret.name,
                 EnumSet.of(CredentialsType.NETRC_FILE)
             )
 
@@ -192,20 +336,21 @@ class InfrastructureServiceServiceIntegrationTest : WordSpec({
 
     "getForId" should {
         "return an infrastructure service" {
-            val userSecret = createOrganizationSecret("user")
-            val passSecret = createOrganizationSecret("pass")
-
             val infrastructureService = service.createForId(
                 OrganizationId(organization.id),
                 SERVICE_NAME,
                 SERVICE_URL,
                 SERVICE_DESC,
-                userSecret.name,
-                passSecret.name,
+                orgUserSecret.name,
+                orgPassSecret.name,
                 EnumSet.of(CredentialsType.NETRC_FILE)
             )
 
             service.getForId(OrganizationId(organization.id), SERVICE_NAME) shouldBe infrastructureService
+        }
+
+        "return null for a non-existing infrastructure service" {
+            service.getForId(OrganizationId(organization.id), "non-existing") should beNull()
         }
     }
 
@@ -253,6 +398,242 @@ class InfrastructureServiceServiceIntegrationTest : WordSpec({
             val result = service.listForId(OrganizationId(organization.id), parameters)
 
             result shouldBe ListQueryResult(services, parameters, services.size.toLong())
+        }
+
+        "apply list query parameters" {
+            val expectedServices = (1..8).map { idx ->
+                InfrastructureService(
+                    name = "$SERVICE_NAME$idx",
+                    url = SERVICE_URL,
+                    description = SERVICE_DESC,
+                    usernameSecret = orgUserSecret,
+                    passwordSecret = orgPassSecret,
+                    organization = organization,
+                    product = null,
+                    repository = null,
+                    credentialsTypes = emptySet()
+                )
+            }
+
+            expectedServices.shuffled().forEach {
+                service.createForId(
+                    id = OrganizationId(organization.id),
+                    name = it.name,
+                    url = it.url,
+                    description = it.description,
+                    usernameSecretRef = it.usernameSecret.name,
+                    passwordSecretRef = it.passwordSecret.name,
+                    credentialsTypes = emptySet()
+                )
+            }
+
+            val parameters =
+                ListQueryParameters(sortFields = listOf(OrderField("name", OrderDirection.ASCENDING)), limit = 4)
+            val services = service.listForId(
+                OrganizationId(organization.id),
+                parameters
+            )
+
+            services.data shouldContainExactly expectedServices.take(4)
+        }
+    }
+
+    "listForHierarchy" should {
+        "return all services for the provided IDs" {
+            val repositoryUrl = "https://repo.example.org/test/repo/"
+            val otherOrg = fixtures.createOrganization("anotherOrganization")
+            val otherProduct = fixtures.createProduct("anotherProduct")
+            val otherRepository = fixtures.createRepository(url = "https://example.com/another-repo.git")
+
+            val match1 = createInfrastructureService(
+                "matching1",
+                "${repositoryUrl}repo1",
+                organization = fixtures.organization
+            )
+            val match2 = createInfrastructureService(
+                "matching2",
+                "${repositoryUrl}repo2",
+                product = fixtures.product
+            )
+            val match3 = createInfrastructureService(
+                "matching3",
+                "${repositoryUrl}repo3",
+                organization = fixtures.organization
+            )
+            val match4 = createInfrastructureService(
+                "matching4",
+                "${repositoryUrl}repo4",
+                repository = fixtures.repository
+            )
+
+            val noMatch1 = createInfrastructureService(
+                "non-matching1",
+                organization = otherOrg,
+                usernameSecret = createOrganizationSecret("user", otherOrg.id),
+                passwordSecret = createOrganizationSecret("pass", otherOrg.id)
+            )
+            val noMatch2 = createInfrastructureService(
+                "non-matching2",
+                repositoryUrl,
+                product = otherProduct,
+                usernameSecret = createProductSecret("user", otherProduct.id),
+                passwordSecret = createProductSecret("pass", otherProduct.id)
+            )
+            val noMatch3 = createInfrastructureService(
+                "non-matching3",
+                repositoryUrl,
+                repository = otherRepository,
+                usernameSecret = createRepositorySecret("user", otherRepository.id),
+                passwordSecret = createRepositorySecret("pass", otherRepository.id)
+            )
+
+            listOf(match1, match2, match3, match4, noMatch1, noMatch2, noMatch3).forEach {
+                createInfrastructureService(it)
+            }
+
+            val services = service.listForHierarchy(
+                Hierarchy(fixtures.repository, fixtures.product, fixtures.organization)
+            )
+
+            services shouldContainExactlyInAnyOrder listOf(match1, match2, match3, match4)
+        }
+
+        "throw exception for duplicated name for same organization" {
+            val productService = createInfrastructureService(product = fixtures.product)
+            val orgService = createInfrastructureService(organization = fixtures.organization)
+            val orgService2 = createInfrastructureService(
+                url = "${SERVICE_URL}/other",
+                organization = fixtures.organization
+            )
+
+            shouldThrow<UniqueConstraintException> {
+                listOf(productService, orgService, orgService2).forEach {
+                    createInfrastructureService(it)
+                }
+            }
+        }
+
+        "handle infrastructure services with duplicate URL correctly (repository beats organization)" {
+            val repositoryService = createInfrastructureService(
+                "MyRepositoryLevelService", repository = fixtures.repository
+            )
+            val orgService = createInfrastructureService(
+                "MyOrganizationLevelService", organization = fixtures.organization
+            )
+            val orgService2 = createInfrastructureService(
+                "MyOrganizationLevelService2", "${SERVICE_URL}/other", organization = fixtures.organization
+            )
+
+            listOf(repositoryService, orgService, orgService2).forEach {
+                createInfrastructureService(it)
+            }
+
+            val services = service.listForHierarchy(
+                Hierarchy(fixtures.repository, fixtures.product, fixtures.organization)
+            )
+
+            services shouldContainExactlyInAnyOrder listOf(repositoryService, orgService2)
+        }
+
+        "handle infrastructure services with duplicate URL correctly (repository beats product)" {
+            val repositoryService = createInfrastructureService(
+                "MyRepositoryLevelService", repository = fixtures.repository
+            )
+            val productService = createInfrastructureService(
+                "MyProductLevelService", product = fixtures.product
+            )
+            val productService2 = createInfrastructureService(
+                "MyProductLevelService2", "${SERVICE_URL}/other", product = fixtures.product
+            )
+
+            listOf(repositoryService, productService, productService2).forEach {
+                createInfrastructureService(it)
+            }
+
+            val services = service.listForHierarchy(
+                Hierarchy(fixtures.repository, fixtures.product, fixtures.organization)
+            )
+
+            services shouldContainExactlyInAnyOrder listOf(repositoryService, productService2)
+        }
+    }
+
+    "getOrCreateDeclarationForRun" should {
+        "create a new entity in the database" {
+            val expectedService = createInfrastructureServiceDeclaration()
+            val infrastructureService = service.getOrCreateDeclarationForRun(expectedService, fixtures.ortRun.id)
+
+            infrastructureService shouldBe expectedService
+
+            val runServices = service.listDeclarationsForRun(fixtures.ortRun.id)
+            runServices shouldContainOnly listOf(infrastructureService)
+        }
+
+        "reuse an already existing entity" {
+            val otherRun = fixtures.createOrtRun()
+            val expectedService = createInfrastructureServiceDeclaration()
+            val serviceForOtherRun = service.getOrCreateDeclarationForRun(expectedService, otherRun.id)
+            val serviceForRun = service.getOrCreateDeclarationForRun(expectedService, fixtures.ortRun.id)
+
+            serviceForRun shouldBe expectedService
+            serviceForRun shouldBe serviceForOtherRun
+
+            val runServices = service.listDeclarationsForRun(fixtures.ortRun.id)
+            runServices shouldContainOnly listOf(serviceForRun)
+        }
+
+        "not reuse a service assigned to an organization" {
+            val orgService = createInfrastructureService(organization = fixtures.organization)
+            createInfrastructureService(orgService)
+
+            val runService = createInfrastructureServiceDeclaration()
+            val dbRunService = service.getOrCreateDeclarationForRun(
+                runService,
+                fixtures.ortRun.id
+            )
+
+            dbRunService shouldNotBe orgService
+        }
+
+        "not reuse a service assigned to a product" {
+            val prodService = createInfrastructureService(product = fixtures.product)
+            createInfrastructureService(prodService)
+
+            val runService = createInfrastructureServiceDeclaration()
+            val dbRunService = service.getOrCreateDeclarationForRun(runService, fixtures.ortRun.id)
+
+            dbRunService shouldNotBe prodService
+        }
+
+        "throw exception if the entity name is invalid" {
+            val serviceName = " #servicename! "
+            val newService = createInfrastructureServiceDeclaration(name = serviceName)
+
+            val exception = shouldThrow<ValidationException> {
+                service.getOrCreateDeclarationForRun(newService, fixtures.ortRun.id)
+            }
+
+            exception.message shouldInclude serviceName
+
+            service.listDeclarationsForRun(fixtures.ortRun.id) shouldBe emptyList()
+        }
+    }
+
+    "listDeclarationsForRun" should {
+        "return all services assigned to a run" {
+            val runService1 = createInfrastructureServiceDeclaration(name = "run1")
+            val runService2 = createInfrastructureServiceDeclaration(name = "run2")
+            val orgService = createInfrastructureService(name = "org", organization = fixtures.organization)
+            val prodService = createInfrastructureService(name = "prod", product = fixtures.product)
+
+            createInfrastructureService(orgService)
+            createInfrastructureService(prodService)
+            service.getOrCreateDeclarationForRun(runService1, fixtures.ortRun.id)
+            service.getOrCreateDeclarationForRun(runService2, fixtures.ortRun.id)
+
+            val runServices = service.listDeclarationsForRun(fixtures.ortRun.id)
+
+            runServices shouldContainExactlyInAnyOrder listOf(runService1, runService2)
         }
     }
 })
